+++ apps/kyb-app/src/components/organisms/DynamicUI/StateManager/components/ActionsHandler/hooks/useEventEmitterLogic/helpers/getDispatchableActions.ts
@@ -1,11 +1,11 @@
 import { EngineManager } from '@app/components/organisms/DynamicUI/StateManager/components/ActionsHandler/helpers/engine-manager';
-import { JsonLogicRuleEngine } from '@app/components/organisms/DynamicUI/rule-engines/json-logic.rule-engine';
-import { Action, UIElement } from '@app/domains/collection-flow';
-import { AnyObject } from '@ballerine/ui';
-import { JsonSchemaRuleEngine } from '@app/components/organisms/DynamicUI/rule-engines/json-schema.rule-engine';
 import { UIState } from '@app/components/organisms/DynamicUI/hooks/useUIStateLogic/types';
 import { DocumentsRuleEngine } from '@app/components/organisms/DynamicUI/rule-engines/documents.rule-engine';
 import { JmespathRuleEngine } from '@app/components/organisms/DynamicUI/rule-engines/jmespath.rule-engine';
+import { JsonLogicRuleEngine } from '@app/components/organisms/DynamicUI/rule-engines/json-logic.rule-engine';
+import { JsonSchemaRuleEngine } from '@app/components/organisms/DynamicUI/rule-engines/json-schema.rule-engine';
+import { Action, UIElement } from '@app/domains/collection-flow';
+import { AnyObject } from '@ballerine/ui';
 
 export const getDispatchableActions = (
   context: AnyObject,
@@ -26,7 +26,8 @@ export const getDispatchableActions = (
     return (
       action.dispatchOn?.rules?.length &&
       action.dispatchOn?.rules?.every(
-        rule => engineManager.getEngine(rule.type).test(context, rule, definition, state).isValid,
+        rule =>
+          engineManager.getEngine(rule.type).validate(context, rule, definition, state).isValid,
       )
     );
   });