+++ apps/kyb-app/src/components/organisms/UIRenderer/elements/JSONForm/hocs/withDynamicUIInput/withDynamicUIInput.tsx
@@ -1,16 +1,15 @@
 import { usePageResolverContext } from '@app/components/organisms/DynamicUI/PageResolver/hooks/usePageResolverContext';
-import { findDefinitionByName } from '@app/components/organisms/UIRenderer/elements/JSONForm/helpers/findDefinitionByName';
-import { useUIElementHandlers } from '@app/components/organisms/UIRenderer/hooks/useUIElementHandlers';
-import { RJSFInputProps, AnyObject, RJSFInputAdapter } from '@ballerine/ui';
-import { useCallback, useMemo } from 'react';
-import get from 'lodash/get';
 import { useStateManagerContext } from '@app/components/organisms/DynamicUI/StateManager/components/StateProvider';
-import { useUIElementProps } from '@app/components/organisms/UIRenderer/hooks/useUIElementProps';
-import { UIElement } from '@app/domains/collection-flow';
 import { useDynamicUIContext } from '@app/components/organisms/DynamicUI/hooks/useDynamicUIContext';
+import { findDefinitionByName } from '@app/components/organisms/UIRenderer/elements/JSONForm/helpers/findDefinitionByName';
 import { useUIElementErrors } from '@app/components/organisms/UIRenderer/hooks/useUIElementErrors/useUIElementErrors';
-import { ErrorsList } from '@ballerine/ui';
+import { useUIElementHandlers } from '@app/components/organisms/UIRenderer/hooks/useUIElementHandlers';
+import { useUIElementProps } from '@app/components/organisms/UIRenderer/hooks/useUIElementProps';
 import { useUIElementState } from '@app/components/organisms/UIRenderer/hooks/useUIElementState';
+import { UIElement } from '@app/domains/collection-flow';
+import { AnyObject, ErrorsList, RJSFInputAdapter, RJSFInputProps } from '@ballerine/ui';
+import get from 'lodash/get';
+import { useCallback, useMemo } from 'react';
 
 const findLastDigit = (str: string) => {
   const digitRegex = /_(\d+)_/g;
@@ -55,11 +54,14 @@ export const withDynamicUIInput = (Component: RJSFInputAdapter<any, any>) => {
     }, [name, currentPage]);
 
     const definition = useMemo(() => {
+      const inputIndex = getInputIndex(inputId || '');
+
       return {
         ...baseDefinition,
+        name: inputIndex !== null ? `${baseDefinition.name}[${inputIndex}]` : baseDefinition.name,
         valueDestination: injectIndexToDestinationIfNeeded(
           baseDefinition.valueDestination,
-          getInputIndex(inputId || ''),
+          inputIndex,
         ),
       };
     }, [baseDefinition, inputId]);
@@ -75,7 +77,7 @@ export const withDynamicUIInput = (Component: RJSFInputAdapter<any, any>) => {
       [definition, elementState, setElementState],
     );
 
-    const { disabled } = useUIElementProps(baseDefinition);
+    const { disabled } = useUIElementProps(definition);
 
     const { onChangeHandler } = useUIElementHandlers(definition);