+++ sdks/web-ui-sdk/CHANGELOG.md
@@ -1,5 +1,34 @@
 # web-ui-sdk
 
+## 1.4.17
+
+### Patch Changes
+
+- Bump ver
+- Updated dependencies
+  - @ballerine/common@0.7.20
+
+## 1.4.16
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.19
+
+## 1.4.15
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.18
+
+## 1.4.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.17
+
 ## 1.4.13
 
 ### Patch Changes