+++ sdks/workflow-node-sdk/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @ballerine/workflow-node-sdk
 
+## 0.5.18
+
+### Patch Changes
+
+- Bump ver
+- Updated dependencies
+  - @ballerine/workflow-core@0.5.18
+
+## 0.5.17
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.5.17
+
+## 0.5.16
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.5.16
+
+## 0.5.15
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.5.15
+
 ## 0.5.14
 
 ### Patch Changes